Refactor: Book 관련 API 엔드포인트 변경
