+++ src/main/java/teamssavice/ssavice/book/controller/BookController.java
@@ -6,10 +6,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.web.PageableDefault;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 import teamssavice.ssavice.auth.constants.Role;
 import teamssavice.ssavice.book.constants.BookStatusFilter;
 import teamssavice.ssavice.book.controller.dto.BookResponse;
@@ -19,16 +16,17 @@ import teamssavice.ssavice.book.service.dto.BookModel;
 import teamssavice.ssavice.global.annotation.CurrentId;
 import teamssavice.ssavice.global.annotation.RequireRole;
 import teamssavice.ssavice.global.dto.PageResponse;
+import teamssavice.ssavice.serviceItem.service.dto.ServiceItemCommand;
 
 
 @RestController
-@RequestMapping("/api")
+@RequestMapping("/api/book")
 @RequiredArgsConstructor
 public class BookController {
 
     private final BookService bookService;
 
-    @GetMapping("/user/book")
+    @GetMapping("/user")
     @RequireRole(Role.USER)
     public ResponseEntity<PageResponse<BookResponse.Info>> getMyBooksByStatus(
         @CurrentId Long userId,
@@ -44,7 +42,7 @@ public class BookController {
         return ResponseEntity.ok(PageResponse.from(reponsePage));
     }
 
-    @GetMapping("/user/book/summary")
+    @GetMapping("/user/summary")
     @RequireRole(Role.USER)
     public ResponseEntity<BookResponse.BookSummary> getBookSummary(
         @CurrentId Long userId
@@ -54,28 +52,24 @@ public class BookController {
         return ResponseEntity.ok(BookResponse.BookSummary.from(model));
     }
 
-    @GetMapping("/company/book")
-    @RequireRole(Role.COMPANY)
-    public ResponseEntity<PageResponse<BookResponse.Info>> getMyCompanysBooksByStatus(
-        @CurrentId Long companyId,
-        @PageableDefault(size = 10) Pageable pageable,
-        @RequestParam BookStatusFilter status
+    @PostMapping("/{serviceId}/apply")
+    @RequireRole(Role.USER)
+    public ResponseEntity<BookResponse.Apply> applyServiceItem(
+            @CurrentId Long userId,
+            @PathVariable Long serviceId
     ) {
-        BookCommand.RetrieveByStatus command = BookCommand.RetrieveByStatus.of(companyId, pageable, status);
-
-        Page<BookModel.Info> models = bookService.getMyCompanysBooksByStatus(command);
-        Page<BookResponse.Info> reponsePage = models.map(BookResponse.Info::from);
-
-        return ResponseEntity.ok(PageResponse.from(reponsePage));
+        BookModel.Apply model = bookService.apply(userId, serviceId);
+        return ResponseEntity.ok(BookResponse.Apply.from(model));
     }
 
-    @GetMapping("/company/book/summary")
-    @RequireRole(Role.COMPANY)
-    public ResponseEntity<BookResponse.BookSummary> getCompanyBookSummary(
-        @CurrentId Long companyId
-    ) {
-        BookModel.BookSummary model = bookService.getCompanysBookSummary(companyId);
 
-        return ResponseEntity.ok(BookResponse.BookSummary.from(model));
+    @PostMapping("/{serviceId}/cancel")
+    @RequireRole(Role.USER)
+    public ResponseEntity<Void> cancelParticipation(
+            @CurrentId Long userId,
+            @PathVariable Long serviceId
+    ) {
+        bookService.cancel(ServiceItemCommand.Cancel.of(userId, serviceId));
+        return ResponseEntity.ok().build();
     }
 }
