+++ src/main/java/teamssavice/ssavice/serviceItem/service/ServiceItemService.java
@@ -10,13 +10,10 @@ import teamssavice.ssavice.address.AddressCommand;
 import teamssavice.ssavice.book.entity.Book;
 import teamssavice.ssavice.book.entity.BookStatus;
 import teamssavice.ssavice.book.service.BookReadService;
-import teamssavice.ssavice.book.service.BookWriteService;
-import teamssavice.ssavice.book.service.dto.BookModel;
 import teamssavice.ssavice.company.entity.Company;
 import teamssavice.ssavice.company.service.CompanyReadService;
 import teamssavice.ssavice.global.constants.ErrorCode;
 import teamssavice.ssavice.global.dto.CursorResult;
-import teamssavice.ssavice.global.exception.ConflictException;
 import teamssavice.ssavice.global.exception.ForbiddenException;
 import teamssavice.ssavice.imageresource.entity.ImageResource;
 import teamssavice.ssavice.imageresource.service.ImageReadService;
@@ -30,8 +27,6 @@ import teamssavice.ssavice.serviceItem.constants.ServiceStatus;
 import teamssavice.ssavice.serviceItem.entity.ServiceItem;
 import teamssavice.ssavice.serviceItem.service.dto.ServiceItemCommand;
 import teamssavice.ssavice.serviceItem.service.dto.ServiceItemModel;
-import teamssavice.ssavice.user.entity.Users;
-import teamssavice.ssavice.user.service.UserReadService;
 import teamssavice.ssavice.wish.service.WishReadService;
 
 import java.util.ArrayList;
@@ -47,8 +42,6 @@ public class ServiceItemService {
     private final ServiceItemReadService serviceItemReadService;
     private final ImageReadService imageReadService;
     private final S3Service s3Service;
-    private final UserReadService userReadService;
-    private final BookWriteService bookWriteService;
     private final BookReadService bookReadService;
     private final RegionReadService regionReadService;
     private final RefundService refundService;
@@ -109,30 +102,6 @@ public class ServiceItemService {
         return ServiceItemModel.Detail.from(serviceItem, imageUrls, isLiked, isBooked);
     }
 
-    @Transactional
-    public BookModel.Apply apply(Long userId, Long serviceId) {
-
-        ServiceItem serviceItem = serviceItemReadService.findById(serviceId);
-        Users user = userReadService.findById(userId);
-
-        validateApply(user, serviceItem);
-
-        serviceItem.participate();
-
-        Book book = bookWriteService.apply(user, serviceItem);
-
-        return BookModel.Apply.of(book.getId());
-    }
-
-    private void validateApply(Users user, ServiceItem serviceItem) {
-
-        serviceItem.validateAppliable();
-
-        if (bookReadService.existsByUserAndServiceAndStatusNot(user.getId(), serviceItem.getId(), BookStatus.CANCELED)) {
-            throw new ConflictException(ErrorCode.ALREADY_APPLIED);
-        }
-    }
-
     public Page<ServiceItemModel.Summary> getServiceItemByCompanyAndStatus(ServiceItemCommand.RetrieveByCompanyAndStatus command) {
         Page<ServiceItem> serviceItems = serviceItemReadService.findByCompanyAndStatus(command);
         return serviceItems.map(ServiceItemModel.Summary::from);
@@ -167,24 +136,5 @@ public class ServiceItemService {
         if (!serviceItem.getCompany().getId().equals(companyId)) {
             throw new ForbiddenException(ErrorCode.NOT_SERVICE_OWNER);
         }
-    }
-
-    @Transactional
-    public void cancel(ServiceItemCommand.Cancel command) {
-
-        ServiceItem serviceItem = serviceItemReadService.findById(command.serviceId());
-        Users user = userReadService.findById(command.userId());
-
-        Book book = bookReadService.findFirstByUserIdAndServiceItemIdOrderByCreatedAtDesc(user.getId(), serviceItem.getId());
-
-        // 최소 인원 검증인데 이거는 현재는 못하게 막아놓고 법적인거 조사하면서 따로 수수료 물면서 환불하는 로직으로 전환예정
-        if (serviceItem.isReachedMinimum()) {
-            throw new ConflictException(ErrorCode.AT_MINIMUM_MEMBER_LIMIT);
-        }
-
-        bookWriteService.cancel(book);
-
-        serviceItem.cancelParticipation();
-        refundService.registerRefunds(List.of(book), serviceItem.getPrice(), RefundReason.USER_CANCEL);
     }
 }
